feat: modularize .h .cpp files
