+++ headers/bank_account.h
@@ -1,0 +1,24 @@
+#ifndef BANK_ACCOUNT_H
+#define BANK_ACCOUNT_H
+
+class BankAccount {
+    private:
+        int accountID;
+        double balance;
+    
+    public:
+        BankAccount(int id, double initialBalance); //construtor
+        ~BankAccount(); //destrutor
+
+        int getAccountID();
+
+        double getBalance();
+
+        void credit(double value);
+
+        void debit(double value);
+
+        void transfer(BankAccount destination, double value);
+};
+
+#endif
