+++ src/bank_account.cpp
@@ -1,0 +1,31 @@
+#include <iostream>
+#include "../headers/bank_account.h"
+using namespace std;
+
+BankAccount::BankAccount(int id, double initialBalance) {
+    this->accountID = id;
+    this->balance = initialBalance;
+}
+
+BankAccount::~BankAccount() {}
+
+int BankAccount::getAccountID() {
+    return this->accountID;
+}
+
+double BankAccount::getBalance() {
+    return this->balance;
+}
+
+void BankAccount::credit(double value) {
+    this->balance = this->balance + value;
+}
+
+void BankAccount::debit(double value) {
+    this->balance = this->balance - value;
+}
+
+void BankAccount::transfer(BankAccount destination, double value) {
+    this->debit(value);
+    destination.credit(value);
+}
